+++ e2e/support/commands.js
@@ -1,17 +1,14 @@
 import "./commands/ui/button";
 import "./commands/ui/icon";
 
+import "./commands/api/index";
 import "./commands/api/alert";
-import "./commands/api/question";
-import "./commands/api/dashboard";
 import "./commands/api/dashboardCard";
-import "./commands/api/collection";
 import "./commands/api/moderation";
 import "./commands/api/pulse";
 import "./commands/api/user";
 import "./commands/api/timeline";
 
-import "./commands/api/composite/createQuestionAndDashboard";
 import "./commands/api/composite/createNativeQuestionAndDashboard";
 import "./commands/api/composite/createQuestionAndAddToDashboard";
 import "./commands/api/composite/createDashboardWithQuestions";
