+++ e2e/support/helpers/e2e-collection-helpers.ts
@@ -5,40 +5,7 @@ import {
   getFullName,
   popover,
 } from "e2e/support/helpers";
-import type {
-  Collection,
-  CollectionId,
-  RegularCollectionId,
-} from "metabase-types/api";
-
-export const createCollection = ({
-  name,
-  description = null,
-  parent_id = null,
-  authority_level = null,
-}: {
-  name: string;
-  description?: string | null;
-  parent_id?: RegularCollectionId | null;
-  authority_level?: "official" | null;
-}): Cypress.Chainable<Cypress.Response<Collection>> => {
-  cy.log(`Create a collection: ${name}`);
-
-  return cy.request("POST", "/api/collection", {
-    name,
-    description,
-    parent_id,
-    authority_level,
-  });
-};
-
-export const archiveCollection = (id: CollectionId) => {
-  cy.log(`Archiving a collection with id: ${id}`);
-
-  return cy.request("PUT", `/api/collection/${id}`, {
-    archived: true,
-  });
-};
+import type { CollectionId } from "metabase-types/api";
 
 /**
  * Clicks the "+" icon on the collection page and selects one of the menu options
